+++ src/main/java/com/powsybl/openloadflow/network/LfShunt.java
@@ -6,12 +6,113 @@
  */
 package com.powsybl.openloadflow.network;
 
+import org.apache.commons.lang3.Range;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
 public interface LfShunt extends LfElement {
+
+    Logger LOGGER = LoggerFactory.getLogger(LfShunt.class);
+
+    class Controller {
+
+        private final String id;
+
+        private final List<Double> sectionsB;
+
+        private final List<Double> sectionsG;
+
+        private int position;
+
+        private final double bMagnitude;
+
+        public Controller(String id, List<Double> sectionsB, List<Double> sectionsG, int position) {
+            this.id = Objects.requireNonNull(id);
+            this.sectionsB = Objects.requireNonNull(sectionsB);
+            this.sectionsG = Objects.requireNonNull(sectionsG);
+            this.position = position;
+            double bMin = Math.min(sectionsB.get(0), sectionsB.get(sectionsB.size() - 1));
+            double bMax = Math.max(sectionsB.get(0), sectionsB.get(sectionsB.size() - 1));
+            this.bMagnitude = Math.abs(bMax - bMin);
+        }
+
+        public String getId() {
+            return id;
+        }
+
+        public List<Double> getSectionsB() {
+            return sectionsB;
+        }
+
+        public int getPosition() {
+            return position;
+        }
+
+        public void setPosition(int position) {
+            this.position = position;
+        }
+
+        public double getB() {
+            return sectionsB.get(this.position);
+        }
+
+        public double getG() {
+            return sectionsG.get(this.position);
+        }
+
+        public double getBMagnitude() {
+            return bMagnitude;
+        }
+
+        private Range<Integer> getAllowedPositionRange(AllowedDirection allowedDirection) {
+            switch (allowedDirection) {
+                case INCREASE:
+                    return Range.between(position, sectionsB.size() - 1);
+                case DECREASE:
+                    return Range.between(0, position);
+                case BOTH:
+                    return Range.between(0, sectionsB.size() - 1);
+                default:
+                    throw new IllegalStateException("Unknown direction: " + allowedDirection);
+            }
+        }
+
+        public Optional<Direction> updateSectionB(double deltaB, int maxSectionShift, AllowedDirection allowedDirection) {
+            // an increase allowed direction means that the section could increase.
+            // a decrease allowed direction means that the section could decrease.
+            double newB = getB() + deltaB;
+            Range<Integer> positionRange = getAllowedPositionRange(allowedDirection);
+
+            int oldSection = position;
+            // find section with the closest b value without exceeding the maximum of sections to switch.
+            double smallestDistance = Math.abs(deltaB);
+            for (int p = positionRange.getMinimum(); p <= positionRange.getMaximum(); p++) {
+                if (Math.abs(p - oldSection) > maxSectionShift) {
+                    // we are not allowed in one outer loop run to go further than maxSectionShift sections
+                    continue;
+                }
+                double distance = Math.abs(newB - sectionsB.get(p));
+                if (distance < smallestDistance) {
+                    position = p;
+                    smallestDistance = distance;
+                }
+            }
+
+            boolean hasChanged = position != oldSection;
+            if (hasChanged) {
+                LOGGER.debug("Controller '{}' change section from {} to {}", id, oldSection, position);
+                return Optional.of(position - oldSection > 0 ? Direction.INCREASE : Direction.DECREASE);
+            }
+            return Optional.empty();
+        }
+    }
 
     double getB();
 
@@ -42,4 +143,10 @@ public interface LfShunt extends LfElement {
     void setVoltageControl(ShuntVoltageControl voltageControl);
 
     void reInit();
+
+    List<Controller> getControllers();
+
+    void updateB();
+
+    void updateG();
 }
