+++ src/main/java/com/powsybl/openloadflow/network/impl/LfShuntImpl.java
@@ -13,9 +13,6 @@ import com.powsybl.iidm.network.ShuntCompensatorModel;
 import com.powsybl.iidm.network.ShuntCompensatorNonLinearModel;
 import com.powsybl.openloadflow.network.*;
 import com.powsybl.openloadflow.util.PerUnit;
-import org.apache.commons.lang3.Range;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -25,8 +22,6 @@ import java.util.stream.Collectors;
  * @author Anne Tilloy <anne.tilloy at rte-france.com>
  */
 public class LfShuntImpl extends AbstractElement implements LfShunt {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(LfShuntImpl.class);
 
     private final List<Ref<ShuntCompensator>> shuntCompensatorsRefs;
 
@@ -45,99 +40,6 @@ public class LfShuntImpl extends AbstractElement implements LfShunt {
     private final double zb;
 
     private double g;
-
-    public static class Controller {
-
-        private final String id;
-
-        private final List<Double> sectionsB;
-
-        private final List<Double> sectionsG;
-
-        private int position;
-
-        private final double bMagnitude;
-
-        public Controller(String id, List<Double> sectionsB, List<Double> sectionsG, int position) {
-            this.id = Objects.requireNonNull(id);
-            this.sectionsB = Objects.requireNonNull(sectionsB);
-            this.sectionsG = Objects.requireNonNull(sectionsG);
-            this.position = position;
-            double bMin = Math.min(sectionsB.get(0), sectionsB.get(sectionsB.size() - 1));
-            double bMax = Math.max(sectionsB.get(0), sectionsB.get(sectionsB.size() - 1));
-            this.bMagnitude = Math.abs(bMax - bMin);
-        }
-
-        public String getId() {
-            return id;
-        }
-
-        public List<Double> getSectionsB() {
-            return sectionsB;
-        }
-
-        public int getPosition() {
-            return position;
-        }
-
-        public void setPosition(int position) {
-            this.position = position;
-        }
-
-        public double getB() {
-            return sectionsB.get(this.position);
-        }
-
-        public double getG() {
-            return sectionsG.get(this.position);
-        }
-
-        public double getBMagnitude() {
-            return bMagnitude;
-        }
-
-        private Range<Integer> getAllowedPositionRange(AllowedDirection allowedDirection) {
-            switch (allowedDirection) {
-                case INCREASE:
-                    return Range.between(position, sectionsB.size() - 1);
-                case DECREASE:
-                    return Range.between(0, position);
-                case BOTH:
-                    return Range.between(0, sectionsB.size() - 1);
-                default:
-                    throw new IllegalStateException("Unknown direction: " + allowedDirection);
-            }
-        }
-
-        public Optional<Direction> updateSectionB(double deltaB, int maxSectionShift, AllowedDirection allowedDirection) {
-            // an increase allowed direction means that the section could increase.
-            // a decrease allowed direction means that the section could decrease.
-            double newB = getB() + deltaB;
-            Range<Integer> positionRange = getAllowedPositionRange(allowedDirection);
-
-            int oldSection = position;
-            // find section with the closest b value without exceeding the maximum of sections to switch.
-            double smallestDistance = Math.abs(deltaB);
-            for (int p = positionRange.getMinimum(); p <= positionRange.getMaximum(); p++) {
-                if (Math.abs(p - oldSection) > maxSectionShift) {
-                    // we are not allowed in one outer loop run to go further than maxSectionShift sections
-                    continue;
-                }
-                double distance = Math.abs(newB - sectionsB.get(p));
-                if (distance < smallestDistance) {
-                    position = p;
-                    smallestDistance = distance;
-                }
-            }
-
-            boolean hasChanged = position != oldSection;
-            if (hasChanged) {
-                LOGGER.debug("Controller '{}' change section from {} to {}", id, oldSection, position);
-                return Optional.of(position - oldSection > 0 ? Direction.INCREASE : Direction.DECREASE);
-            }
-            return Optional.empty();
-        }
-    }
 
     public LfShuntImpl(List<ShuntCompensator> shuntCompensators, LfNetwork network, LfBus bus, boolean voltageControlCapability) {
         // if withVoltageControl equals to true, all shunt compensators that are listed must control voltage.
@@ -224,6 +126,7 @@ public class LfShuntImpl extends AbstractElement implements LfShunt {
         this.b = b;
     }
 
+    @Override
     public void updateB() {
         setB(controllers.stream().mapToDouble(Controller::getB).sum());
     }
@@ -233,6 +136,7 @@ public class LfShuntImpl extends AbstractElement implements LfShunt {
         return g;
     }
 
+    @Override
     public void updateG() {
         setG(controllers.stream().mapToDouble(Controller::getG).sum());
     }
